+++ packages/protocol/test/TaikoL1Oracle.t.sol
@@ -375,17 +375,15 @@ contract TaikoL1OracleTest is TaikoL1TestBase {
         uint32 gasUsed = 1_000_000;
 
         for (
-            uint256 blockId = 1;
-            blockId < conf.blockMaxProposals * 10;
-            blockId++
+            uint64 blockId = 1; blockId < conf.blockMaxProposals * 10; blockId++
         ) {
             TaikoData.BlockMetadata memory meta =
                 proposeBlock(Alice, 1_000_000, 1024);
             printVariables("after propose");
             mine(1);
 
-            bytes32 blockHash = bytes32(1e10 + blockId);
-            bytes32 signalRoot = bytes32(1e9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + uint256(blockId));
+            bytes32 signalRoot = bytes32(1e9 + uint256(blockId));
 
             proveBlock(
                 Dave,
